Improved in-memory feed meta-data modification. MILESTONE UPDATE: Additional details ...
The changes here largely stabalize the new direction for the rss management screen. This marks the full transition from immediately updating the database to making changes to a cached copy that is later applied to the database. This sequence of operations will prevent constant thrashing of the database state while separate thread operations are active against the database on the main UI screen.

FEED METADATA MODIFICATION
- Maintain existing feed metadata link when matching input is re-introduced.
-- Keep the selection and modify the correct entry when both keys (name and url) change.
- Clear entry from cache an reapply to cache as a safeguard against stale entries.

RESET BUTTON
- Toggle reset button's text based on edit mode.
-- A new entry will have the reset button show the text = "Clear". That is a more appropriate description. You are clearing text you've entered.
-- An existing entry has the reset button show the text = "New" since that exits existing entry edit mode. You can then add an entry.
-- Reset sequence matches the approach used in the selection handler when updating the name and url fields. GTK Entry Buffers directly modified.

ROW SELECTION
- More efficient use of selected numeric id instead of url text.
- Reworked selection loop. Previously, the first item in the list was skipped.

Cleaned up diagnostic messages.

Other bugs exist ... however, this is a milestone update.