RSS CLI - Lib updates
Release Status
--------------------------------------------------------------------------------
Version 7.3: This supports a production ready release. This release is the first production ready release of the RSS program. The cli program is for the Terminal / command-line situations.

The primary operation of the cli program is to perform RSS feed download through the command-line. This can be beneficial when a UI is not desired. This cli program will download all feeds automatically or ones specifically specified on the command-line. The cli program is production ready but not part of the production release since it needs updates to work with the same database as the GUI program. It has diverged from the GUI program in that it currently modifies a database separate from the GUI program. The cli program will eventually get updated to operate as a command-line way to manage the same database used by the GUI.

Module: rss cli (command-line operations).
--------------------------------------------------------------------------------
RSS development began with a program like this. A command-line program. This is the most effective way to build out code foundations for a GUI program.

Changes
--------------------------------------------------------------------------------
- 1. Code compatible with the RSS GUI
-- Numerous changes were made to the RSS libraries used by the RSS GUI. This version of the RSS cli has been updated to make function calls with the same API function signatures as those used by the RSS GUI.
--- Impact: Reliable use of the cli.
- 2. Build scripts
-- Makefiles were updated to use the same standards as for the GUI. Some modifications however were done to lessen a few warnings due to the way argtable is defined.
-- Impact: Consistent build process.