[Enhancement/Optimization] Remove old articles on program start-up.
An earlier feature shown in the RSS management screen but never implemented. Modeled an approach to slim down the database when a feed's retention period is 1 or 7 days. The program will remove all feed articles that match the expired time frame.

The addition of automatic clean-up improves program speed for typical use. The database will not grow too big and the amount of data loaded into RAM will often be on the low side.

The program has been tested for worse case scenarios involving a greater accumulation of data. The speed optimizations over the last month has an even greater impact when applied to typical case with automatic clean-up.

Additional changes in this commit involve database access.

- SQL statements are formatted in a more consistent way.
- More extensive SQL statement for creating tables.
- Added indexes to boost look up speed. A controversial change. The database is not expected to have enough data to justify maintenance of an index. Added it anyway.
- Added unique constraints on the tables as an additional data integrity safeguard. Now impossible to violate fundamental requirements on natural keys (feed name and feed url).