[Polish] Unselect rows API.
The unselect row function is used to change a row's status from [selected] to [unselected]. The program uses this function when a GTKNotebook tab/page is switched from one tab to another.

Out of the box, a GTKNotebook has the functionality to show the contents of a tab when that tab is clicked. That is all GTKNotebook does. The programmer has to add interactive functionality that applies to the contents. In this case, a side panel exists where the data shown in that side panel is based on the row selected in the tab. The goal is to [remember] what was selected so when a tab that was clicked in the past is clicked again the future, whatever row was selected would be selected again. When that earlier row is once again selected, the information in the side panel should update accordingly.

A list box has a [selection signal] that it emits when a row is selected. Once that signal goes out, the programmer can tie that signal to functionality. The problem is that the list box widget in GTK is [smart] enough to know if a row it currently has selected has already [sent a signal]. That means you can't just select the same row again to [re-emit] the signal.

Clearing the row selection also resets the list box's memory of what was selected. That way, the same row can be selected twice in order to re-emit a signal. The unselect row function is then used on a given tab's list box to clear it out so that when the tab is clicked again, the same row that was selected before can be selected again to trigger the chain of events that will result in the side panel getting updated.