[Optimization] Removed unused code.
A pair of functions to clear out article data was used by the application. A variable exists in app_win.cpp that tracks which article is the one being accessed by the user. Previously, when a row was selected in a tab representing a feed, that row's data would be copied into the article variable. That ensured that the same article information was accessed across functions. A good practice was to clear out the data fields associated with that variable.

The new change is a list of selected articles exist. Accessing the list means a completed article object is retrieved that can be used to update the article variable. That eliminates the need to clear out the variable and so the clear functions are no longer needed nor are there any anticipated needs for such a function. This reduces the amount of code resident in the program.