[Polish] Maintain row selection.
Each tab represents an RSS feed. The contents of the tab is a GTK List Box. The list box for the feed represents the headlines for the feed. What you end up with is:

-- feed [tab]
-------- headlines [list box]

A side panel exists that shows the extended description for a headline. When a headline is selected from the [list box] that keys into other information that is then shown on the side panel. You end up with something similar to this:

[feed 1 name] ...... [feed 2 name] ...... [feed 3 name] .... | Side Panel [selected feed name]
[feed 1 headlines] . [feed 2 headlines] . [feed 3 headlines] | .... [headline title / headline details]

When an item from a feed is selected, the details are applied to the side panel. GTK does not know anything about this. All GTK provides is the API and building blocks for you to setup the flow and output.

When a headline is selected in the form of a list box, all the list box retains is that a row was selected. The programming functionality is to detect when a row is selected and apply the information associated with that row to another visual item. In this case, a side panel.

The problem is that GTK remembers which row is selected on a list box. You cannot select the row twice and expect detection to occur again. As you switch tabs, the row that was selected for the list box associated with that tab remains in effect. However, the side panel does not get updated.

This change improves tracking of which row was selected for a tabs [feed] list box [headlines]. The information is then resequenced to update the side panel accordingly.

Additional Details:
--------------------------------------------------------------------------------
Persistent row selection is now consistent in the UI. The following commits are the foundation for this latest change. Full details related to this update exist in these commits.

- Improve row select. Setup row clear: 83e8f9de557c03f6284b088cf468597b17cb74b0
- Prior row activated from tab switch: 3503c00797c715213a0ca47b897a943f4f2914e3
- Select row by url: 2a530b0179382a102e8021f4ff12b6682dbacb15