[Production Release 7.6] Build for Fedora / Ubuntu and Microsoft Windows.
Microsoft Windows compatibility has been achieved. This release is the first release where the compiled program will run under Microsoft Windows.

The build was modified to subtract Microsoft Windows specific code when compiling under Linux and the reverse when cross compiling to Microsoft Windows. The MinGW-64 compiler and MinGW-64 libraries were used to achieve the translation under a Linux host to product Windows compatible binaries.  The Windows version was *minimally* tested and the code was minimally adjusted as needed to stabilize the Windows version. The project remains oriented towards a Linux/Unix and the Microsoft Windows compatibility test validates the code readiness for future portability.