[Code readability] Indent top-level scope block in implementation file.
The commits referenced below resulted in several statements gathered into scope/linkage declaration blocks defined in .cpp implementation files. The content of these linkage controlling blocks is indented accordingly. See the following commits for details on the changes. The describe changes for internal and external linkage.

[Statement Organization] Reduced need for static declarations in UI code.
Used unnamed namespaces to contain code in .cpp implementation files where the intent is for code to have internal linkage. Alleviates the need to mark the enclosed code as static. The practical result is the code in the implementation file has an enclosing scope similar to a singleton OOP class. All enclosed members are only visible in the implementation file. In some cases, a closure is generated that realizes a form of RAII when the referenced symbols are no longer accessed in an execution path at run time.

The combination of named namespaces in header files in which the declarations have external linkage that are defined in implementation files but with unnamed namespaces that provide further implementation with internal linkage makes this functional design a first-class way to approach encapsulation with exported interfaces.

I came upon this solution while reviewing reading materials related to C++17. I decided this is an approach compatible with the source code design. After putting it into practice, I saw that this provides a cleaner way to separate interface from implementation in a manner that allows declarations/definitions of different linkage categories to exist in the same file. This is very clean and for additional visibility guarantees, each declaration can still be marked static. I did not elect that option in the final code definition. However, I tested it and it works either way. The other advantage here is that the number of forward function declarations can be reduced. For now, I find it more consistent to retain declarations for each defined function.

See the following commits:
- ec56911a64f83edf55c55af9d7e4d9750beb8f56
- 148813e3c5bb9fd6c4f6c77fc9d9ba0c1b5546e2
- 1df33e1b1546af612aa5997aac5cd5f6d1dc55ab

[Statement Organization] Grouped declarations with external C linkage.
All code declarations with external C linkage is now easier to identify. Grouping the code in a single area for this purpose makes it easier to review the code; easier to add new declarations of the same type; and speeds readability when attempting to locate callback functions that need to be reviewed.

See the following commit:
- 913888f738d58ffe0faeff9cd97f4f13190cf205