[C++17 string_view] Removed reference qualifier.
When using a std::string_view, a reference qualifier is not needed. In some cases, a reference qualifier can cause issues if the byte encoding of the strings change. That issue did not occur in the implementation of the indicates_html function but was observed elsewhere. That understanding was applied here as a preventative measure.