[DB/XML] *Major improvement* Enhanced interpretation of parsed XML before input into SQL database.
Websites vary in the quality of their implemenation of RSS/Atam feeds. Not all websites present RSS/Atom feed information in a clean or complete way. I previously wrote basic mitigations in app_win.cpp in the user interface. However, that was always a stop gap measure to make the program functional. This update provides a true translation layer that checks and further formats data parsed from an XML stream before insertion of the translated data into an SQL database.

Parts of the data are now keyed to appropriate fields designating which data is plain text vs html. This simplifies decisions in the user interface in terms which areas of the screen to assign data based on a data field's assumed format. Data from websites that do not meet minimum standards for the RSS/Atom schema is rejected to prevent polluting the database and to eliminate the need to add further mitigations.

The changes were tested on a small, sample dataset. The results were beyond satisfactory. The translation function finalize_rss_article implements the process. It provides core verification of the parsed XML data and formats information as needed. The final result is reflected in the passed instance of rss_article.