[Reliability Improvement] Verify and condition downloaded data.
Data that is downloaded and parsed is checked for issues and reconditioned if necessary to meet the requirements of the program. That includes swapping data field values to correspond to their intended use. The process looks for plain-text vs. HTML encoded data. Specific fields are expected to have data in a particular format.

The process that transfers data downloaded from the Internet into an SQL database calls a function that ensure data is located in the correct fields based on the encoding detected in data fields. Data not meeting minimum requirements for entry to the database is rejected to avoid awkward or useless output in the user interface. Since data format issues are not user serviceable (the data is read-only from the Internet), rejecting the data is the best approach in this instance.