[Usability] Optimized status messages.
The main goal of this release was to achieve 3 things:

 1.] Fix issues in RSS/Atom parsing.
 --- Some feeds do not supply all data according to the RSS standards. The program adapts to those situations primarily by rejecting incomplete feed data.

 2.] Improve performance. Refine the code. This is more of a polishing effort.
---- The program itself has been at an acceptable performance level since release 7.2 (see commit 006cb188103e48cf4cc0fbb9b8a058a5e9485a95).
Now that C++17 is mature on most compilers and with widespread availability of C++20 on the horizon, it was time to move more of the code out of a C++11/C++14 posture and use some of the capabilities in C++17 to improve performance and reduce code.
C++17 string_view was useful in that the program processes a fairly sizeable amount of text in a limited time window. The string_view was successfully used during XML parsing to curtail string copies. Along the way, some gotchas with string_view was observed which triggered other code improvements.
C++17 structured bindings are now used with range-for loops. The downside is unused temporaries are generated in some cases. Since this program sub-divides larger ranges into smaller ones to iterate, the observed performance impact is negligible, but worth noting for other situations that may arise in other programs.
C++17 unnamed namespaces are very useful in this code as it provides a streamlined and clearly identifiable way to define implementation file scoped definitions.

 3.] Make headline rendering more algorithmic for purposes of reusability.
---- The code in app_win.cpp for initially building the tab contents, the displayed headlines, has been made more reusable. Since the download process also adds headlines to the screen, it had code that was very similar to that which ran when the program first started up. I saw a chance to consolidate these processes in terms of algorithmic expression. The process was refined until the only difference between the processes that accumulate headlines onto the screen was the data source they used for that purpose.
The algorithm changes improved performance. The change in data structure from a multi-valued tuple to a string also reduced RAM consumption.
Storage reads were significantly reduced. Many unnecessary storage reads occured in the prior release and this release not only reduces RAM but greatly reduces storage seeks (SSD or HDD).

Other Changes:

 - Single use function reverted to inline statements. The function was get_tab_contents_container_by_feed_name. The function was not clear in its definition with respect to its inputs and outputs. The better decision was to revert it to inline statements given that it was used once or twice in the program.
 -- This commit removes 1 usage of get_tab_contents_container_by_feed_name and places the contents of that function inline in the function.
    The other usage is described in commit add7e26ac8d0c32c31f243e62969d83ed5a1b0ad/

 - Status Message Improvements
 -- The final piece of polish in this release is that status messages. Too many status messages can reduce code efficiency. Too few status messages can foster uncertainty. The goal was to achieve the right level of verbosity presenting the status of the program.
 -- The prior release did well showing status during download but after downloads were done, the status messages were less potent or meaningful. The program now has a more complete status disclosure path that works well during downloads as well as when navigating tabs and headlines.

General Note:

This is intended to be the final source code modification for Gautier RSS Reader 7. What that means is this version of Gautier RSS has reached a sufficient level of reliability, performance, and feature definition that it is useful without further modification.

Keep in mind:

 - This does not mean this version is perfect. Many programs can be improved almost indefinitely.
 - This version is the best version of an RSS/Atom reader I have ever made.
 - I have not lost interest in maintaining the RSS reader. My original intent was to create an RSS reader I could successfully use.

The reason this is the final modification, save a serious bug, is all the software development so far has been to get the program to a reliable, solid, efficiently running state that is trustworthy and where there are no lingering issues that tug at me. I wanted to get all the issues addressed before I froze series 7.
The other reason this is the final modification for version 7 is I want this version to exist as my version of a good entry level C++ desktop program. This program has the following aspects that I want to record:

 - Represent a careful intersection of solid Modern C++ and Modern C as I understood it in early 2020.
 -- Given an "application developer" viewpoint, the program reflects a careful selection of elements culled from the following standards:
 --- C++17. Many of the additions for C++17 were observed from the book, Modern C++ Programming Cookbook by Marius Bancila as well as cppreference.com.
 --- C Interfaces and Implementations by David R. Hanson. The overall structure and approach in this program is influenced by this paperback book suggested by Dani Richards.
 --- Best selection of C APIs unified under a C++17 program. C++ syntax, tools, and definitions would be used to integrate C APIs into an overall C++ program. The C APIs include:
 ---- Curl
 ---- SQLite
 ---- GNOME XML
 ---- GNOME Toolkit (aka GTK 3)
 ---- GNU C library for accessing UNIX/Linux file systems

*** With all that, a simple and clean RSS reader now exists. I will be using this reader for the next several years.
At the same time, I will begin a project in few months that will start on a version 8. Series 8 will use GTK 4, C++20, future Linux standards and will be ready in a few years.