Incorporate user interface distribution component.
Tracking the embedded resource C implementation file generated by glib2-generate-resource tool. Not a new file for the build but is being tracked in git for the first time. Although it is useful to regenerate the file source/rss_ui/app_resources.c, it is helpful to have the file read to go when cloning the repository in a new environment. The app_icon.png file is an exact copy of newsreader.png file. The files are used as part of the FreeDesktop manifest xml.