Removed webkit_settings_set_enable_plugins.
Latest documentation for GTK+ 3 says this is deprecated since since version 2.32. Bug report https://bugs.webkit.org/show_bug.cgi?id=220620. Good discussion. As of 1/14/2021, a change in how use of this function is flagged in the GTK library. I interpret the discussion to conclude that the function is not applicable or implemented and is safe to remove. It is uncertain if the absence of this function indicates plugins are disabled by default but it is not critical at this point to address this.